feat: use optimistic UI, first query cache and then overwrite with network data
This is done without merging the elements from followup queries in the same array, but instead letting the queries being saved in cache in different entries for each combination of variables (in this case {"page":"2"} for example). That way, when we change pages, we put the data we already have in the cache for that variables as a quick optimistic response and then launch the query to the server that will overwrite the cache results.